move test ecosystem within crate
