+++ .github/workflows/rust.yml
@@ -28,7 +28,7 @@ jobs:
     - name: Build (Default)
       run: cargo build
     - name: Test (Default)
-      run: cargo test --features tests -- --nocapture
+      run: cargo test -- --nocapture
     - name: Build (All features)
       run: cargo build --all-features
     - name: Test (All features)
