+++ rinex/src/tests/mod.rs
@@ -1,0 +1,17 @@
+//! rinex lib test modules
+pub mod toolkit;
+
+mod antex;
+mod clocks;
+mod compression;
+mod decompression;
+mod ionex;
+mod merge;
+mod meteo;
+mod nav;
+mod obs;
+mod parsing;
+mod production;
+mod sampling;
+mod smoothing;
+mod stats;
