+++ rinex/src/tests/stats.rs
@@ -1,7 +1,7 @@
 #[cfg(test)]
 mod test {
-    use rinex::*;
-    use rinex::{observation::*, prelude::*};
+    use crate::*;
+    use crate::{observation::*, prelude::*};
     use std::str::FromStr;
     fn run_test(rinex: &Rinex, ops: &str, expected: Vec<(Sv, Vec<(Observable, f64)>)>) {
         let (clk_stats, stats) = match ops {
